Build tables strictly with ORM
